% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_priors.R
\name{.compact_dag_str}
\alias{.compact_dag_str}
\title{Compact a dagitty object to a single-line string}
\usage{
.compact_dag_str(dag)
}
\description{
Compact a dagitty object to a single-line string
}
\keyword{internal}
