% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_causalqueries.R
\name{.dagitty_to_cq_string}
\alias{.dagitty_to_cq_string}
\title{Convert a dagitty DAG to a CausalQueries model string}
\usage{
.dagitty_to_cq_string(dag)
}
\arguments{
\item{dag}{A dagitty object}
}
\value{
Character string, e.g. "X -> Y; Z -> Y"
}
\description{
Convert a dagitty DAG to a CausalQueries model string
}
\keyword{internal}
