% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{.elicit_ui}
\alias{.elicit_ui}
\title{UI for one elicitation panel (used by both Pre- and Post-Study tabs)}
\usage{
.elicit_ui(id, dag_str, mode, locked_dag = FALSE)
}
\arguments{
\item{id}{Module namespace id (\code{"pre"} or \code{"post"}).}

\item{dag_str}{Initial DAG specification string.}

\item{mode}{Initial elicitation mode.}
}
\description{
UI for one elicitation panel (used by both Pre- and Post-Study tabs)
}
\keyword{internal}
