% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_priors.R
\name{.from_csv}
\alias{.from_csv}
\title{Read a structured CSV exported by \code{export_priors_csv}}
\usage{
.from_csv(file)
}
\description{
Read a structured CSV exported by \code{export_priors_csv}
}
\keyword{internal}
