% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_utils.R
\name{format_query}
\alias{format_query}
\title{Format a conditional probability query as a readable string}
\usage{
format_query(node, parent_values)
}
\arguments{
\item{node}{Character. Name of the outcome node}

\item{parent_values}{Named integer vector of parent node values (may be
length-0 for root nodes)}
}
\value{
Character string of the form "P(node = 1 | ...)"
}
\description{
Format a conditional probability query as a readable string
}
\keyword{internal}
