% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_utils.R
\docType{data}
\name{LIKERT_LABELS}
\alias{LIKERT_LABELS}
\title{Display lines for the Likert scale}
\format{
An object of class \code{character} of length 7.
}
\usage{
LIKERT_LABELS
}
\description{
Display lines for the Likert scale
}
\keyword{internal}
