% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_utils.R
\docType{data}
\name{LIKERT_PROBS}
\alias{LIKERT_PROBS}
\title{Named numeric vector mapping 7-point Likert keys to probabilities}
\format{
An object of class \code{numeric} of length 7.
}
\usage{
LIKERT_PROBS
}
\description{
Named numeric vector mapping 7-point Likert keys to probabilities
}
\keyword{internal}
