% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit_cpt.R
\name{print.cpt}
\alias{print.cpt}
\title{Print a conditional probability table}
\usage{
\method{print}{cpt}(x, digits = 4L, ...)
}
\arguments{
\item{x}{A \code{cpt} object}

\item{digits}{Integer. Decimal places for probabilities}

\item{...}{Ignored}
}
\description{
Print a conditional probability table
}
