% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{summary.elicit_dag_result}
\alias{summary.elicit_dag_result}
\title{Summary method for elicit_dag_result}
\usage{
\method{summary}{elicit_dag_result}(object, ...)
}
\arguments{
\item{object}{An \code{elicit_dag_result} object}

\item{...}{Passed to \code{print.elicit_dag_result}}
}
\description{
Delegates to \code{\link{print.elicit_dag_result}}.
}
